pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "textoverlay"
version = "0.1.0"
description = "Intelligent text overlay placement and rendering (programmatic API)"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "TextOverlay Contributors" }
]
license = { text = "MIT" }
keywords = ["computer-vision", "image-processing", "graphics", "text-overlay", "cv"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "opencv-python>=4.8.0",
  "pillow>=9.0.0",
  "numpy>=1.21.0"
]

[project.optional-dependencies]
# Saliency and feature extraction extras
saliency = [
  "torch>=1.10",
  "torchvision>=0.11"
]

[tool.setuptools]
packages = ["textoverlay"]
include-package-data = true

[tool.setuptools.package-data]
textoverlay = []
